+++ examples/custom-runtime-in-extract-mode/webpack.config.js
@@ -1,12 +1,7 @@
-/* eslint-disable import/no-extraneous-dependencies */
 const path = require('path');
 const merge = require('deepmerge');
-const TextExtractPlugin = require('extract-text-webpack-plugin');
 const baseConfig = require('../base-webpack.config');
 const SpritePlugin = require('../../plugin');
-
-const CSSExtractor = new TextExtractPlugin('[name].css');
-const HTMLExtractor = new TextExtractPlugin('[name].html');
 
 const config = merge(baseConfig, {
   context: __dirname,
@@ -24,23 +19,12 @@ const config = merge(baseConfig, {
         loader: 'svg-sprite-loader',
         options: {
           extract: true,
-          spriteFilename: 'sprite-[hash:6].svg'
+          runtimeGenerator: require.resolve('./extracting-runtime-generator.js')
         }
-      },
-      {
-        test: /\.css$/,
-        loader: CSSExtractor.extract({ use: 'css-loader' })
-      },
-      {
-        test: /\.html$/,
-        loader: HTMLExtractor.extract({ use: 'html-loader' })
       }
     ]
   },
-
   plugins: [
-    CSSExtractor,
-    HTMLExtractor,
     new SpritePlugin()
   ]
 });