+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.1.0-3"></a>
+# [2.1.0-3](https://github.com/kisenka/webpack-svg-sprite-loader/compare/v2.1.0-2...v2.1.0-3) (2017-05-10)
+
+
+### Bug Fixes
+
+* **plugin:** properly replace paths on Windows ([0c70caa](https://github.com/kisenka/webpack-svg-sprite-loader/commit/0c70caa)), closes [#106](https://github.com/kisenka/webpack-svg-sprite-loader/issues/106)
+
+
+
+<a name="2.1.0-2"></a>
+# [2.1.0-2](https://github.com/kisenka/webpack-svg-sprite-loader/compare/v2.1.0-1...v2.1.0-2) (2017-05-09)
+
+
+### Bug Fixes
+
+* **loader:** symbol id interpolation ([18edd99](https://github.com/kisenka/webpack-svg-sprite-loader/commit/18edd99))
+
+
+
+<a name="2.1.0-1"></a>
+# [2.1.0-1](https://github.com/kisenka/webpack-svg-sprite-loader/compare/v2.1.0-0...v2.1.0-1) (2017-05-08)
+
+
+### Bug Fixes
+
+* **plugin:** properly generate symbol url in extract mode ([6af7230](https://github.com/kisenka/webpack-svg-sprite-loader/commit/6af7230))
+
+
+
+<a name="2.1.0-0"></a>
+# [2.1.0-0](https://github.com/kisenka/webpack-svg-sprite-loader/compare/v2.0.5...v2.1.0-0) (2017-05-07)
+
+
+### Bug Fixes
+
+* **utils:** fix default import ([0c34daa](https://github.com/kisenka/webpack-svg-sprite-loader/commit/0c34daa))
+
+
+### Features
+
+* **interop:** extract-text-webpack-plugin & html-webpack-plugin interop ([a38fdcc](https://github.com/kisenka/webpack-svg-sprite-loader/commit/a38fdcc))
+* **interop:** extract-text-webpack-plugin with allChunks: true interoperability ([63d347d](https://github.com/kisenka/webpack-svg-sprite-loader/commit/63d347d))
+* **spritehash:** add ability to use `[spritehash]` substitution token in spriteFilename ([f9eba1b](https://github.com/kisenka/webpack-svg-sprite-loader/commit/f9eba1b))
+
+
+
 <a name="2.0.6"></a>
 ## [2.0.6](https://github.com/kisenka/webpack-svg-sprite-loader/compare/v2.0.5...v2.0.6) (2017-05-13)